move logger and io_pool to core
